apps build but database collapses due to not update version of it

+++ app/src/main/java/com/example/taskscheduler/data/TaskTracking.kt
@@ -25,7 +25,7 @@ import java.util.Date
 @TypeConverters(DateConverters::class)
 data class TaskTracking(
     @PrimaryKey
-    val taskId: Long, // same id as Task
+    val taskId: Int, // same id as Task
 
     var timesCompleted: Int = 0,
     var totalTimeMillisSpent: Long = 0,
